+++ crates/ruma-events/src/room/message/server_notice.rs
@@ -6,7 +6,6 @@ use crate::PrivOwnedStr;
 /// The payload for a server notice message.
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-#[serde(tag = "msgtype", rename = "m.server_notice")]
 pub struct ServerNoticeMessageEventContent {
     /// A human-readable description of the notice.
     pub body: String,