+++ crates/ruma-events/src/room/message/video.rs
@@ -13,7 +13,6 @@ use crate::room::{
 /// The payload for a video message.
 #[derive(Clone, Debug, Deserialize, Serialize)]
 #[cfg_attr(not(ruma_unstable_exhaustive_types), non_exhaustive)]
-#[serde(tag = "msgtype", rename = "m.video")]
 pub struct VideoMessageEventContent {
     /// A description of the video.
     ///